@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: metalstack/io/accounting/api/v1/cluster.proto
 
@@ -669,7 +669,7 @@ func file_metalstack_io_accounting_api_v1_cluster_proto_rawDescGZIP() []byte {
 }
 
 var file_metalstack_io_accounting_api_v1_cluster_proto_msgTypes = make([]protoimpl.MessageInfo, 6)
-var file_metalstack_io_accounting_api_v1_cluster_proto_goTypes = []interface{}{
+var file_metalstack_io_accounting_api_v1_cluster_proto_goTypes = []any{
 	(*ClusterReport)(nil),            // 0: metalstack.io.accounting.api.v1.ClusterReport
 	(*ClusterWorkerGroup)(nil),       // 1: metalstack.io.accounting.api.v1.ClusterWorkerGroup
 	(*ClusterUsageRequest)(nil),      // 2: metalstack.io.accounting.api.v1.ClusterUsageRequest
@@ -719,7 +719,7 @@ func file_metalstack_io_accounting_api_v1_cluster_proto_init() {
 	}
 	file_metalstack_io_accounting_api_v1_common_proto_init()
 	if !protoimpl.UnsafeEnabled {
-		file_metalstack_io_accounting_api_v1_cluster_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_cluster_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*ClusterReport); i {
 			case 0:
 				return &v.state
@@ -731,7 +731,7 @@ func file_metalstack_io_accounting_api_v1_cluster_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_cluster_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_cluster_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*ClusterWorkerGroup); i {
 			case 0:
 				return &v.state
@@ -743,7 +743,7 @@ func file_metalstack_io_accounting_api_v1_cluster_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_cluster_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_cluster_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*ClusterUsageRequest); i {
 			case 0:
 				return &v.state
@@ -755,7 +755,7 @@ func file_metalstack_io_accounting_api_v1_cluster_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_cluster_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_cluster_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*ClusterUsageResponse); i {
 			case 0:
 				return &v.state
@@ -767,7 +767,7 @@ func file_metalstack_io_accounting_api_v1_cluster_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_cluster_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_cluster_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*ClusterUsage); i {
 			case 0:
 				return &v.state
@@ -779,7 +779,7 @@ func file_metalstack_io_accounting_api_v1_cluster_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_cluster_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_cluster_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*ClusterUsageAccumuluated); i {
 			case 0:
 				return &v.state
