@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: metalstack/io/accounting/api/v1/common.proto
 
@@ -377,7 +377,7 @@ func file_metalstack_io_accounting_api_v1_common_proto_rawDescGZIP() []byte {
 }
 
 var file_metalstack_io_accounting_api_v1_common_proto_msgTypes = make([]protoimpl.MessageInfo, 4)
-var file_metalstack_io_accounting_api_v1_common_proto_goTypes = []interface{}{
+var file_metalstack_io_accounting_api_v1_common_proto_goTypes = []any{
 	(*Empty)(nil),                    // 0: metalstack.io.accounting.api.v1.Empty
 	(*Report)(nil),                   // 1: metalstack.io.accounting.api.v1.Report
 	(*KubernetesResourceReport)(nil), // 2: metalstack.io.accounting.api.v1.KubernetesResourceReport
@@ -404,7 +404,7 @@ func file_metalstack_io_accounting_api_v1_common_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_metalstack_io_accounting_api_v1_common_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_common_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Empty); i {
 			case 0:
 				return &v.state
@@ -416,7 +416,7 @@ func file_metalstack_io_accounting_api_v1_common_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_common_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_common_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*Report); i {
 			case 0:
 				return &v.state
@@ -428,7 +428,7 @@ func file_metalstack_io_accounting_api_v1_common_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_common_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_common_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*KubernetesResourceReport); i {
 			case 0:
 				return &v.state
@@ -440,7 +440,7 @@ func file_metalstack_io_accounting_api_v1_common_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_common_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_common_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*UsageQuery); i {
 			case 0:
 				return &v.state
@@ -453,7 +453,7 @@ func file_metalstack_io_accounting_api_v1_common_proto_init() {
 			}
 		}
 	}
-	file_metalstack_io_accounting_api_v1_common_proto_msgTypes[3].OneofWrappers = []interface{}{}
+	file_metalstack_io_accounting_api_v1_common_proto_msgTypes[3].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
