@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: metalstack/io/accounting/api/v1/info.proto
 
@@ -389,7 +389,7 @@ func file_metalstack_io_accounting_api_v1_info_proto_rawDescGZIP() []byte {
 }
 
 var file_metalstack_io_accounting_api_v1_info_proto_msgTypes = make([]protoimpl.MessageInfo, 6)
-var file_metalstack_io_accounting_api_v1_info_proto_goTypes = []interface{}{
+var file_metalstack_io_accounting_api_v1_info_proto_goTypes = []any{
 	(*TimeQuery)(nil),             // 0: metalstack.io.accounting.api.v1.TimeQuery
 	(*Project)(nil),               // 1: metalstack.io.accounting.api.v1.Project
 	(*ProjectInfoRequest)(nil),    // 2: metalstack.io.accounting.api.v1.ProjectInfoRequest
@@ -421,7 +421,7 @@ func file_metalstack_io_accounting_api_v1_info_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_metalstack_io_accounting_api_v1_info_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_info_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*TimeQuery); i {
 			case 0:
 				return &v.state
@@ -433,7 +433,7 @@ func file_metalstack_io_accounting_api_v1_info_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_info_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_info_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*Project); i {
 			case 0:
 				return &v.state
@@ -445,7 +445,7 @@ func file_metalstack_io_accounting_api_v1_info_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_info_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_info_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*ProjectInfoRequest); i {
 			case 0:
 				return &v.state
@@ -457,7 +457,7 @@ func file_metalstack_io_accounting_api_v1_info_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_info_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_info_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*ProjectInfoResponse); i {
 			case 0:
 				return &v.state
@@ -469,7 +469,7 @@ func file_metalstack_io_accounting_api_v1_info_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_info_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_info_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*TenantInfoRequest); i {
 			case 0:
 				return &v.state
@@ -481,7 +481,7 @@ func file_metalstack_io_accounting_api_v1_info_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_info_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_info_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*TenantInfoResponse); i {
 			case 0:
 				return &v.state
