@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: metalstack/io/accounting/api/v1/ip.proto
 
@@ -667,7 +667,7 @@ func file_metalstack_io_accounting_api_v1_ip_proto_rawDescGZIP() []byte {
 }
 
 var file_metalstack_io_accounting_api_v1_ip_proto_msgTypes = make([]protoimpl.MessageInfo, 6)
-var file_metalstack_io_accounting_api_v1_ip_proto_goTypes = []interface{}{
+var file_metalstack_io_accounting_api_v1_ip_proto_goTypes = []any{
 	(*IP)(nil),                    // 0: metalstack.io.accounting.api.v1.IP
 	(*IPReport)(nil),              // 1: metalstack.io.accounting.api.v1.IPReport
 	(*IPUsageRequest)(nil),        // 2: metalstack.io.accounting.api.v1.IPUsageRequest
@@ -715,7 +715,7 @@ func file_metalstack_io_accounting_api_v1_ip_proto_init() {
 	}
 	file_metalstack_io_accounting_api_v1_common_proto_init()
 	if !protoimpl.UnsafeEnabled {
-		file_metalstack_io_accounting_api_v1_ip_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_ip_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*IP); i {
 			case 0:
 				return &v.state
@@ -727,7 +727,7 @@ func file_metalstack_io_accounting_api_v1_ip_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_ip_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_ip_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*IPReport); i {
 			case 0:
 				return &v.state
@@ -739,7 +739,7 @@ func file_metalstack_io_accounting_api_v1_ip_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_ip_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_ip_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*IPUsageRequest); i {
 			case 0:
 				return &v.state
@@ -751,7 +751,7 @@ func file_metalstack_io_accounting_api_v1_ip_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_ip_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_ip_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*IPUsageResponse); i {
 			case 0:
 				return &v.state
@@ -763,7 +763,7 @@ func file_metalstack_io_accounting_api_v1_ip_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_ip_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_ip_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*IPUsage); i {
 			case 0:
 				return &v.state
@@ -775,7 +775,7 @@ func file_metalstack_io_accounting_api_v1_ip_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_ip_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_ip_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*IPUsageAccumuluated); i {
 			case 0:
 				return &v.state
@@ -788,7 +788,7 @@ func file_metalstack_io_accounting_api_v1_ip_proto_init() {
 			}
 		}
 	}
-	file_metalstack_io_accounting_api_v1_ip_proto_msgTypes[2].OneofWrappers = []interface{}{}
+	file_metalstack_io_accounting_api_v1_ip_proto_msgTypes[2].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
