@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: metalstack/io/accounting/api/v1/machine.proto
 
@@ -656,7 +656,7 @@ func file_metalstack_io_accounting_api_v1_machine_proto_rawDescGZIP() []byte {
 }
 
 var file_metalstack_io_accounting_api_v1_machine_proto_msgTypes = make([]protoimpl.MessageInfo, 5)
-var file_metalstack_io_accounting_api_v1_machine_proto_goTypes = []interface{}{
+var file_metalstack_io_accounting_api_v1_machine_proto_goTypes = []any{
 	(*MachineReport)(nil),            // 0: metalstack.io.accounting.api.v1.MachineReport
 	(*MachineUsageRequest)(nil),      // 1: metalstack.io.accounting.api.v1.MachineUsageRequest
 	(*MachineUsageResponse)(nil),     // 2: metalstack.io.accounting.api.v1.MachineUsageResponse
@@ -703,7 +703,7 @@ func file_metalstack_io_accounting_api_v1_machine_proto_init() {
 	}
 	file_metalstack_io_accounting_api_v1_common_proto_init()
 	if !protoimpl.UnsafeEnabled {
-		file_metalstack_io_accounting_api_v1_machine_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_machine_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*MachineReport); i {
 			case 0:
 				return &v.state
@@ -715,7 +715,7 @@ func file_metalstack_io_accounting_api_v1_machine_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_machine_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_machine_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*MachineUsageRequest); i {
 			case 0:
 				return &v.state
@@ -727,7 +727,7 @@ func file_metalstack_io_accounting_api_v1_machine_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_machine_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_machine_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*MachineUsageResponse); i {
 			case 0:
 				return &v.state
@@ -739,7 +739,7 @@ func file_metalstack_io_accounting_api_v1_machine_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_machine_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_machine_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*MachineUsage); i {
 			case 0:
 				return &v.state
@@ -751,7 +751,7 @@ func file_metalstack_io_accounting_api_v1_machine_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_machine_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_machine_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*MachineUsageAccumuluated); i {
 			case 0:
 				return &v.state
@@ -764,7 +764,7 @@ func file_metalstack_io_accounting_api_v1_machine_proto_init() {
 			}
 		}
 	}
-	file_metalstack_io_accounting_api_v1_machine_proto_msgTypes[1].OneofWrappers = []interface{}{}
+	file_metalstack_io_accounting_api_v1_machine_proto_msgTypes[1].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
