@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: metalstack/io/accounting/api/v1/network_traffic.proto
 
@@ -650,7 +650,7 @@ func file_metalstack_io_accounting_api_v1_network_traffic_proto_rawDescGZIP() []
 }
 
 var file_metalstack_io_accounting_api_v1_network_traffic_proto_msgTypes = make([]protoimpl.MessageInfo, 6)
-var file_metalstack_io_accounting_api_v1_network_traffic_proto_goTypes = []interface{}{
+var file_metalstack_io_accounting_api_v1_network_traffic_proto_goTypes = []any{
 	(*NetworkTraffic)(nil),           // 0: metalstack.io.accounting.api.v1.NetworkTraffic
 	(*NetworkTrafficReport)(nil),     // 1: metalstack.io.accounting.api.v1.NetworkTrafficReport
 	(*NetworkUsageRequest)(nil),      // 2: metalstack.io.accounting.api.v1.NetworkUsageRequest
@@ -691,7 +691,7 @@ func file_metalstack_io_accounting_api_v1_network_traffic_proto_init() {
 	}
 	file_metalstack_io_accounting_api_v1_common_proto_init()
 	if !protoimpl.UnsafeEnabled {
-		file_metalstack_io_accounting_api_v1_network_traffic_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_network_traffic_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*NetworkTraffic); i {
 			case 0:
 				return &v.state
@@ -703,7 +703,7 @@ func file_metalstack_io_accounting_api_v1_network_traffic_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_network_traffic_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_network_traffic_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*NetworkTrafficReport); i {
 			case 0:
 				return &v.state
@@ -715,7 +715,7 @@ func file_metalstack_io_accounting_api_v1_network_traffic_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_network_traffic_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_network_traffic_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*NetworkUsageRequest); i {
 			case 0:
 				return &v.state
@@ -727,7 +727,7 @@ func file_metalstack_io_accounting_api_v1_network_traffic_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_network_traffic_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_network_traffic_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*NetworkUsageResponse); i {
 			case 0:
 				return &v.state
@@ -739,7 +739,7 @@ func file_metalstack_io_accounting_api_v1_network_traffic_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_network_traffic_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_network_traffic_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*NetworkUsage); i {
 			case 0:
 				return &v.state
@@ -751,7 +751,7 @@ func file_metalstack_io_accounting_api_v1_network_traffic_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_network_traffic_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_network_traffic_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*NetworkUsageAccumuluated); i {
 			case 0:
 				return &v.state
@@ -764,7 +764,7 @@ func file_metalstack_io_accounting_api_v1_network_traffic_proto_init() {
 			}
 		}
 	}
-	file_metalstack_io_accounting_api_v1_network_traffic_proto_msgTypes[2].OneofWrappers = []interface{}{}
+	file_metalstack_io_accounting_api_v1_network_traffic_proto_msgTypes[2].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
