@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: metalstack/io/accounting/api/v1/pod.proto
 
@@ -746,7 +746,7 @@ func file_metalstack_io_accounting_api_v1_pod_proto_rawDescGZIP() []byte {
 }
 
 var file_metalstack_io_accounting_api_v1_pod_proto_msgTypes = make([]protoimpl.MessageInfo, 6)
-var file_metalstack_io_accounting_api_v1_pod_proto_goTypes = []interface{}{
+var file_metalstack_io_accounting_api_v1_pod_proto_goTypes = []any{
 	(*Container)(nil),                  // 0: metalstack.io.accounting.api.v1.Container
 	(*PodReport)(nil),                  // 1: metalstack.io.accounting.api.v1.PodReport
 	(*ContainerUsageRequest)(nil),      // 2: metalstack.io.accounting.api.v1.ContainerUsageRequest
@@ -793,7 +793,7 @@ func file_metalstack_io_accounting_api_v1_pod_proto_init() {
 	}
 	file_metalstack_io_accounting_api_v1_common_proto_init()
 	if !protoimpl.UnsafeEnabled {
-		file_metalstack_io_accounting_api_v1_pod_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_pod_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Container); i {
 			case 0:
 				return &v.state
@@ -805,7 +805,7 @@ func file_metalstack_io_accounting_api_v1_pod_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_pod_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_pod_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*PodReport); i {
 			case 0:
 				return &v.state
@@ -817,7 +817,7 @@ func file_metalstack_io_accounting_api_v1_pod_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_pod_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_pod_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*ContainerUsageRequest); i {
 			case 0:
 				return &v.state
@@ -829,7 +829,7 @@ func file_metalstack_io_accounting_api_v1_pod_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_pod_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_pod_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*ContainerUsageResponse); i {
 			case 0:
 				return &v.state
@@ -841,7 +841,7 @@ func file_metalstack_io_accounting_api_v1_pod_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_pod_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_pod_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*ContainerUsage); i {
 			case 0:
 				return &v.state
@@ -853,7 +853,7 @@ func file_metalstack_io_accounting_api_v1_pod_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_pod_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_pod_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*ContainerUsageAccumuluated); i {
 			case 0:
 				return &v.state
@@ -866,7 +866,7 @@ func file_metalstack_io_accounting_api_v1_pod_proto_init() {
 			}
 		}
 	}
-	file_metalstack_io_accounting_api_v1_pod_proto_msgTypes[2].OneofWrappers = []interface{}{}
+	file_metalstack_io_accounting_api_v1_pod_proto_msgTypes[2].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
