@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: metalstack/io/accounting/api/v1/postgres.proto
 
@@ -813,7 +813,7 @@ func file_metalstack_io_accounting_api_v1_postgres_proto_rawDescGZIP() []byte {
 }
 
 var file_metalstack_io_accounting_api_v1_postgres_proto_msgTypes = make([]protoimpl.MessageInfo, 6)
-var file_metalstack_io_accounting_api_v1_postgres_proto_goTypes = []interface{}{
+var file_metalstack_io_accounting_api_v1_postgres_proto_goTypes = []any{
 	(*Postgres)(nil),                  // 0: metalstack.io.accounting.api.v1.Postgres
 	(*PostgresReport)(nil),            // 1: metalstack.io.accounting.api.v1.PostgresReport
 	(*PostgresUsageRequest)(nil),      // 2: metalstack.io.accounting.api.v1.PostgresUsageRequest
@@ -861,7 +861,7 @@ func file_metalstack_io_accounting_api_v1_postgres_proto_init() {
 	}
 	file_metalstack_io_accounting_api_v1_common_proto_init()
 	if !protoimpl.UnsafeEnabled {
-		file_metalstack_io_accounting_api_v1_postgres_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_postgres_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Postgres); i {
 			case 0:
 				return &v.state
@@ -873,7 +873,7 @@ func file_metalstack_io_accounting_api_v1_postgres_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_postgres_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_postgres_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*PostgresReport); i {
 			case 0:
 				return &v.state
@@ -885,7 +885,7 @@ func file_metalstack_io_accounting_api_v1_postgres_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_postgres_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_postgres_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*PostgresUsageRequest); i {
 			case 0:
 				return &v.state
@@ -897,7 +897,7 @@ func file_metalstack_io_accounting_api_v1_postgres_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_postgres_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_postgres_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*PostgresUsageResponse); i {
 			case 0:
 				return &v.state
@@ -909,7 +909,7 @@ func file_metalstack_io_accounting_api_v1_postgres_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_postgres_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_postgres_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*PostgresUsage); i {
 			case 0:
 				return &v.state
@@ -921,7 +921,7 @@ func file_metalstack_io_accounting_api_v1_postgres_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_postgres_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_postgres_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*PostgresUsageAccumuluated); i {
 			case 0:
 				return &v.state
@@ -934,7 +934,7 @@ func file_metalstack_io_accounting_api_v1_postgres_proto_init() {
 			}
 		}
 	}
-	file_metalstack_io_accounting_api_v1_postgres_proto_msgTypes[2].OneofWrappers = []interface{}{}
+	file_metalstack_io_accounting_api_v1_postgres_proto_msgTypes[2].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
