@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: metalstack/io/accounting/api/v1/product_option.proto
 
@@ -591,7 +591,7 @@ func file_metalstack_io_accounting_api_v1_product_option_proto_rawDescGZIP() []b
 
 var file_metalstack_io_accounting_api_v1_product_option_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_metalstack_io_accounting_api_v1_product_option_proto_msgTypes = make([]protoimpl.MessageInfo, 5)
-var file_metalstack_io_accounting_api_v1_product_option_proto_goTypes = []interface{}{
+var file_metalstack_io_accounting_api_v1_product_option_proto_goTypes = []any{
 	(ProductOption)(0),                     // 0: metalstack.io.accounting.api.v1.ProductOption
 	(*ProductOptionReport)(nil),            // 1: metalstack.io.accounting.api.v1.ProductOptionReport
 	(*ProductOptionUsageRequest)(nil),      // 2: metalstack.io.accounting.api.v1.ProductOptionUsageRequest
@@ -634,7 +634,7 @@ func file_metalstack_io_accounting_api_v1_product_option_proto_init() {
 	}
 	file_metalstack_io_accounting_api_v1_common_proto_init()
 	if !protoimpl.UnsafeEnabled {
-		file_metalstack_io_accounting_api_v1_product_option_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_product_option_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*ProductOptionReport); i {
 			case 0:
 				return &v.state
@@ -646,7 +646,7 @@ func file_metalstack_io_accounting_api_v1_product_option_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_product_option_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_product_option_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*ProductOptionUsageRequest); i {
 			case 0:
 				return &v.state
@@ -658,7 +658,7 @@ func file_metalstack_io_accounting_api_v1_product_option_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_product_option_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_product_option_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*ProductOptionUsageResponse); i {
 			case 0:
 				return &v.state
@@ -670,7 +670,7 @@ func file_metalstack_io_accounting_api_v1_product_option_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_product_option_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_product_option_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*ProductOptionUsage); i {
 			case 0:
 				return &v.state
@@ -682,7 +682,7 @@ func file_metalstack_io_accounting_api_v1_product_option_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_product_option_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_product_option_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*ProductOptionUsageAccumuluated); i {
 			case 0:
 				return &v.state
@@ -695,7 +695,7 @@ func file_metalstack_io_accounting_api_v1_product_option_proto_init() {
 			}
 		}
 	}
-	file_metalstack_io_accounting_api_v1_product_option_proto_msgTypes[1].OneofWrappers = []interface{}{}
+	file_metalstack_io_accounting_api_v1_product_option_proto_msgTypes[1].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
