@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: metalstack/io/accounting/api/v1/s3.proto
 
@@ -742,7 +742,7 @@ func file_metalstack_io_accounting_api_v1_s3_proto_rawDescGZIP() []byte {
 }
 
 var file_metalstack_io_accounting_api_v1_s3_proto_msgTypes = make([]protoimpl.MessageInfo, 6)
-var file_metalstack_io_accounting_api_v1_s3_proto_goTypes = []interface{}{
+var file_metalstack_io_accounting_api_v1_s3_proto_goTypes = []any{
 	(*S3Bucket)(nil),              // 0: metalstack.io.accounting.api.v1.S3Bucket
 	(*S3BucketReport)(nil),        // 1: metalstack.io.accounting.api.v1.S3BucketReport
 	(*S3UsageRequest)(nil),        // 2: metalstack.io.accounting.api.v1.S3UsageRequest
@@ -785,7 +785,7 @@ func file_metalstack_io_accounting_api_v1_s3_proto_init() {
 	}
 	file_metalstack_io_accounting_api_v1_common_proto_init()
 	if !protoimpl.UnsafeEnabled {
-		file_metalstack_io_accounting_api_v1_s3_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_s3_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*S3Bucket); i {
 			case 0:
 				return &v.state
@@ -797,7 +797,7 @@ func file_metalstack_io_accounting_api_v1_s3_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_s3_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_s3_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*S3BucketReport); i {
 			case 0:
 				return &v.state
@@ -809,7 +809,7 @@ func file_metalstack_io_accounting_api_v1_s3_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_s3_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_s3_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*S3UsageRequest); i {
 			case 0:
 				return &v.state
@@ -821,7 +821,7 @@ func file_metalstack_io_accounting_api_v1_s3_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_s3_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_s3_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*S3UsageResponse); i {
 			case 0:
 				return &v.state
@@ -833,7 +833,7 @@ func file_metalstack_io_accounting_api_v1_s3_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_s3_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_s3_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*S3Usage); i {
 			case 0:
 				return &v.state
@@ -845,7 +845,7 @@ func file_metalstack_io_accounting_api_v1_s3_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_s3_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_s3_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*S3UsageAccumuluated); i {
 			case 0:
 				return &v.state
@@ -858,7 +858,7 @@ func file_metalstack_io_accounting_api_v1_s3_proto_init() {
 			}
 		}
 	}
-	file_metalstack_io_accounting_api_v1_s3_proto_msgTypes[2].OneofWrappers = []interface{}{}
+	file_metalstack_io_accounting_api_v1_s3_proto_msgTypes[2].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
