@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: metalstack/io/accounting/api/v1/volume.proto
 
@@ -721,7 +721,7 @@ func file_metalstack_io_accounting_api_v1_volume_proto_rawDescGZIP() []byte {
 }
 
 var file_metalstack_io_accounting_api_v1_volume_proto_msgTypes = make([]protoimpl.MessageInfo, 6)
-var file_metalstack_io_accounting_api_v1_volume_proto_goTypes = []interface{}{
+var file_metalstack_io_accounting_api_v1_volume_proto_goTypes = []any{
 	(*Volume)(nil),                   // 0: metalstack.io.accounting.api.v1.Volume
 	(*VolumeReport)(nil),             // 1: metalstack.io.accounting.api.v1.VolumeReport
 	(*VolumeUsageRequest)(nil),       // 2: metalstack.io.accounting.api.v1.VolumeUsageRequest
@@ -768,7 +768,7 @@ func file_metalstack_io_accounting_api_v1_volume_proto_init() {
 	}
 	file_metalstack_io_accounting_api_v1_common_proto_init()
 	if !protoimpl.UnsafeEnabled {
-		file_metalstack_io_accounting_api_v1_volume_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_volume_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Volume); i {
 			case 0:
 				return &v.state
@@ -780,7 +780,7 @@ func file_metalstack_io_accounting_api_v1_volume_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_volume_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_volume_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*VolumeReport); i {
 			case 0:
 				return &v.state
@@ -792,7 +792,7 @@ func file_metalstack_io_accounting_api_v1_volume_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_volume_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_volume_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*VolumeUsageRequest); i {
 			case 0:
 				return &v.state
@@ -804,7 +804,7 @@ func file_metalstack_io_accounting_api_v1_volume_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_volume_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_volume_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*VolumeUsageResponse); i {
 			case 0:
 				return &v.state
@@ -816,7 +816,7 @@ func file_metalstack_io_accounting_api_v1_volume_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_volume_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_volume_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*VolumeUsage); i {
 			case 0:
 				return &v.state
@@ -828,7 +828,7 @@ func file_metalstack_io_accounting_api_v1_volume_proto_init() {
 				return nil
 			}
 		}
-		file_metalstack_io_accounting_api_v1_volume_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_metalstack_io_accounting_api_v1_volume_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*VolumeUsageAccumuluated); i {
 			case 0:
 				return &v.state
@@ -841,7 +841,7 @@ func file_metalstack_io_accounting_api_v1_volume_proto_init() {
 			}
 		}
 	}
-	file_metalstack_io_accounting_api_v1_volume_proto_msgTypes[2].OneofWrappers = []interface{}{}
+	file_metalstack_io_accounting_api_v1_volume_proto_msgTypes[2].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
